+++ src/parsers.js
@@ -265,11 +265,24 @@ const parseDomainWhois = (domain, whois) => {
 		.split('\n')
 		.map((line) => line.replace('\t', '  '))
 
-
 	// Parse WHOIS info for specific TLDs
 
-	if (domain.endsWith('.uk') || domain.endsWith('.be') || domain.endsWith('.nl') || domain.endsWith('.eu') || domain.endsWith('.ly') || domain.endsWith('.mx')) {
+	if (
+		domain.endsWith('.uk') ||
+		domain.endsWith('.be') ||
+		domain.endsWith('.nl') ||
+		domain.endsWith('.eu') ||
+		domain.endsWith('.ly') ||
+		domain.endsWith('.mx') ||
+		domain.endsWith('.gg') ||
+		domain.endsWith('.je') ||
+		domain.endsWith('.as')
+	) {
 		lines = handleMultiLines(lines)
+	}
+
+	if (domain.endsWith('.gg') || domain.endsWith('.je') || domain.endsWith('.as')) {
+		lines = handleMissingColons(lines)
 	}
 
 	if (domain.endsWith('.ua')) {
@@ -307,14 +320,12 @@ const parseDomainWhois = (domain, whois) => {
 			if (data[label] && Array.isArray(data[label])) {
 				data[label].push(value)
 			} else if (!ignoreLabels.includes(label.toLowerCase()) && !ignoreTexts.some((text) => label.toLowerCase().includes(text))) {
-
 				// WHOIS field already exists, if so append data
 				if (data[label] && data[label] !== value) {
 					data[label] = `${data[label]} ${value}`.trim()
 				} else {
 					data[label] = value
 				}
-
 			} else {
 				text.push(line)
 			}
@@ -411,18 +422,30 @@ const handleJpLines = (lines) => {
 			line = line.replace(/^[a-z]. \[/, '[')
 		}
 
-		if (line.startsWith("[ ")) {
+		if (line.startsWith('[ ')) {
 			// skip
-		} else if (line.startsWith("[")) {
+		} else if (line.startsWith('[')) {
 			ret.push(line)
-		} else if (line.startsWith(" ")) {
+		} else if (line.startsWith(' ')) {
 			const prev = ret.pop()
-			ret.push(prev + "\n" + line.trim())
+			ret.push(prev + '\n' + line.trim())
 		} else {
 			// skip
 		}
 	}
 	return ret.map((line) => line.replace(/\[(.*?)\]/g, '$1:'))
+}
+
+// Handle formats like this:
+// Registrar Gandi SAS
+const handleMissingColons = (lines) => {
+	lines.forEach((line, index) => {
+		if (line.startsWith('Registrar ')) {
+			lines[index] = line.replace('Registrar ', 'Registrar: ')
+		}
+	})
+
+	return lines
 }
 
 module.exports.parseSimpleWhois = parseSimpleWhois